# common-platform-reference-master-data-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-reference-master-data
  description: Reference Master Data Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/master-data'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/countries/{countryId}':
    get:
      tags:
        - Countries
      summary: Get Country by ID
      description: Retrieve a specific country by its ID.
      operationId: getCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Country retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        '404':
          description: Country not found
    put:
      tags:
        - Countries
      summary: Update Country
      description: Update an existing country by its ID.
      operationId: updateCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
        required: true
      responses:
        '200':
          description: Country updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        '404':
          description: Country not found
    delete:
      tags:
        - Countries
      summary: Delete Country
      description: Delete a specific country by its ID.
      operationId: deleteCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Country deleted successfully
        '404':
          description: Country not found
...
  /api/v1/countries/filter:
    post:
      tags:
        - Countries
      summary: List Countries
      description: Retrieve a paginated list of countries.
      operationId: filterCountries
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestCountryDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved list of countries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseCountryDTO'
  /api/v1/countries:
    post:
      tags:
        - Countries
      summary: Create Country
      description: Create a new country.
      operationId: createCountry
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountryDTO'
        required: true
      responses:
        '200':
          description: Country created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDTO'
        '400':
          description: Invalid input data
...
  '/api/v1/legal-forms/country/{countryId}':
    get:
      tags:
        - Legal Forms
      summary: Get Legal Forms by Country
      description: Retrieve all legal forms for a specific country.
      operationId: getLegalFormsByCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved legal forms for the country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LegalFormDTO'
  '/api/v1/identity-documents/country/{countryId}':
    get:
      tags:
        - Identity Documents
      summary: List Identity Documents by Country
      description: Retrieve a paginated list of identity documents for a specific country.
      operationId: listIdentityDocumentsByCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: 'The direction of sorting, either ASC or DESC.'
          required: false
          schema:
            type: string
            description: 'The direction of sorting, either ASC or DESC.'
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: >-
            Successfully retrieved list of identity documents for the specified
            country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
  '/api/v1/activity-codes/country/{countryId}':
    get:
      tags:
        - Activity Codes
      summary: Get Activity Codes by Country
      description: Retrieve all activity codes for a specific country.
      operationId: getActivityCodesByCountry
      parameters:
        - name: countryId
          in: path
          description: ID of the country
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved activity codes for the country
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityCodeDTO'
components:
  schemas:
...
    LegalFormDTO:
      type: object
      properties:
        legalFormId:
          type: integer
          format: int64
        countryId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        isCorporate:
          type: boolean
        entityType:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
...
    CountryDTO:
      type: object
      properties:
        countryId:
          type: integer
          format: int64
          readOnly: true
        isoCode:
          type: string
        countryName:
          type: string
        region:
          type: string
          enum:
            - EUROPE
            - AMERICAS
            - APAC
            - OCEANIA
            - ASIA
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
        svgFlag:
          type: string
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
...
    ActivityCodeDTO:
      type: object
      properties:
        activityCodeId:
          type: integer
          format: int64
        countryId:
          type: integer
          format: int64
        code:
          type: string
        classificationSys:
          type: string
        description:
          type: string
        parentCodeId:
          type: integer
          format: int64
        highRisk:
          type: boolean
        riskFactors:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestCountryDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CountryDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponseCountryDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CountryDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.